requirements.txt
fastapi[standard]==0.115.4
pyee==11.0.1
sqlalchemy==2.0.36
pgvector==0.3.6
langchain==0.3
langchain_openai==0.2.6
python-dotenv==1.0.1
pydash==8.0.3
asyncpg==0.29.0